+++ edx_proctoring/views.py
@@ -523,8 +523,24 @@ class StudentOnboardingStatusByCourseView(ProctoredAPIView):
                 data['status'] = InstructorDashboardOnboardingAttemptStatus.other_course_approved
                 data['modified'] = other_verified_attempt.modified
             else:
-                data['status'] = (InstructorDashboardOnboardingAttemptStatus
-                                  .get_onboarding_status_from_attempt_status(user_attempt.get('status')))
+                attempt_status = user_attempt.get('status')
+
+                # If the learner's most recent attempt is in the "onboarding_reset" state,
+                # return the onboarding_reset_past_due state.
+                # This is a consequence of a bug in our software that allows a learner to end up
+                # with their only or their most recent exam attempt being in the "onboarding_reset" state.
+                # The learner should not end up in this state, but while we work on a fix, we should not
+                # display "null" in the Instructor Dashboard Student Onboarding Panel.
+                # TODO: remove as part of MST-745
+                if user_attempt.get('status') == ProctoredExamStudentAttemptStatus.onboarding_reset:
+                    onboarding_status = InstructorDashboardOnboardingAttemptStatus.onboarding_reset_past_due
+                else:
+                    onboarding_status = \
+                        InstructorDashboardOnboardingAttemptStatus.get_onboarding_status_from_attempt_status(
+                            attempt_status
+                        )
+
+                data['status'] = onboarding_status
                 data['modified'] = user_attempt.get('modified')
 
             onboarding_data.append(data)